test(coverage): rspec and simplecov setup

+++ db/migrate/20210925215325_create_deputies.rb
@@ -3,7 +3,7 @@ class CreateDeputies < ActiveRecord::Migration[6.0]
     create_table :deputies do |t|
       t.string :cpf, null: false, unique: true
       t.bigint :ide, null: false, unique: true
-      t.integer :parlamentary_card, null: false, unique: true
+      t.bigint :parlamentary_card, null: false, unique: true
       t.string :name, null: false
       t.string :state, null: false
